+++ src/main/java/svenhjol/charm/base/CharmCompat.java
@@ -1,17 +1,19 @@
 package svenhjol.charm.base;
 
+import svenhjol.charm.Charm;
+import svenhjol.charm.base.compat.QuarkCompat;
 import svenhjol.meson.MesonInstance;
+import svenhjol.meson.helper.ForgeHelper;
 
 public class CharmCompat {
     public static void init(MesonInstance instance) {
-        // TODO quark 1.15
-//        try {
-//            if (ForgeHelper.isModLoaded("quark")) {
-//                Charm.quarkCompat = QuarkCompat.class.newInstance();
-//                instance.log.debug("Loaded Quark compatibility");
-//            }
-//        } catch (Exception e) {
-//            instance.log.error("Error loading Quark compatibility: " + e.getMessage());
-//        }
+        try {
+            if (ForgeHelper.isModLoaded("quark")) {
+                Charm.quarkCompat = QuarkCompat.class.newInstance();
+                instance.log.debug("Loaded Quark compatibility");
+            }
+        } catch (Exception e) {
+            instance.log.error("Error loading Quark compatibility: " + e.getMessage());
+        }
     }
 }