+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.5.7
 * Added configurable parrot mimic chance. Fixes #245
+* Added cauldron cleaning things.
+* Updated some textures from MCVinnyq
 
 ## 1.5.6
 * Added parrots perching on nearby end rods.